requirements.txt
anthropic==0.45.2
blessed==1.20.0
prompt_toolkit==3.0.50
pydantic==2.10.6
pygame==2.6.1
PyYAML==6.0.2
rich==13.9.4
